Show Send button when no Swaps
